A. robetta

+++ src/base/ProtocolMessageHeader.java
@@ -2,7 +2,7 @@ package base;
 
 public class ProtocolMessageHeader {
     public int type;
-    public long data_len;
+    public int data_len;
     public static class Types {
         public final static int _message = -3;
         public final static int _logout = -2;
@@ -11,7 +11,7 @@ public class ProtocolMessageHeader {
         public final static int _int = 1;
         public final static int _long = 2;
     }
-    public ProtocolMessageHeader(int type, long data_len)
+    public ProtocolMessageHeader(int type, int data_len)
     {
         this.type = type;
         this.data_len = data_len;
